Move config.json to app/ directory, remove community prop, update imports
