+++ app/components/organisms/navbar/navbar.styles.tsx
@@ -1,6 +1,8 @@
-import styled from 'styled-components';
+import styled, {css} from 'styled-components';
 
-const desktopMediaQuery = '@media (min-width: 768px)';
+const desktopMediaQuery = css`
+  ${({theme}) => theme.mediaQueries.desktop}
+`;
 
 interface NavBarLinkProps {
   active?: boolean | string;
