+++ app/components/templates/community/community.styles.tsx
@@ -1,6 +1,8 @@
-import styled from 'styled-components';
+import styled, {css} from 'styled-components';
 
-const desktopMediaQuery = '@media (min-width: 768px)';
+const desktopMediaQuery = css`
+  ${({theme}) => theme.mediaQueries.desktop}
+`;
 
 export const CommunitySection = styled.section`
   display: flex;
